% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multicontracted}
\alias{multicontracted}
\title{multicontracted}
\usage{
multicontracted(graphlist, attrib, pdbreak, vertex.size.attrib = "degree",
  vertex.color.attrib = "degdiff", color.attrib = "degdiff",
  vertex.label.minmax = c(0.2, 0.5), vertex.size.minmax = c(20, 35),
  edge.width.minmax = c(3, 6), edge.label.minmax = c(0.01, 0.01),
  savepng = F, savename = "mutiContractedGraph", png.width = 15,
  png.height = 12, png.res = 500)
}
\arguments{
\item{graphlist}{a list of igraph objects (i.e., graphs)}

\item{attrib}{character name of igraph vector attribute for coloring, sizing}

\item{pdbreak}{a vector of the numeric values to divide periods}

\item{vertex.size.attrib}{character {'degree','indeg','outdeg','degdiff'}}

\item{veretex.color.attrib}{character {'degree','indeg','outdeg','degdiff',
<other attribute>}}

\item{color.attrib}{character specifying attribute for vertex color scale}

\item{vertex.label.minmax}{numeric vector of length 2: minmax vertex label}

\item{vertex.size.minmax}{numeric vector of length 2: minmax vertex size}

\item{edge.width.minmax}{numeric vector of length 2: minmax edge width}

\item{edge.label.minmax}{numeric vector of length 2: minmax edge label}

\item{savepng}{logical saveplot as png}

\item{savename}{character name to save plot to working directory}

\item{png.width}{numeric width in inches of png output}

\item{png.height}{numeric height in inches of png output}

\item{png.res}{numeric resolution of png out}
}
\value{
list of length two: 1. gout: subgraphs by time period for each
igraph object in list; 2. namedf: df of attribute names and values for
the subgraph nodes.
}
\description{
Multiframe plot of igraph objects
}
\details{
This function creates a multiframe of subgraphs from a list of
igraph objects. Each graph in the list is a row and each time period is
a column. Each igraph object in the list must have the same vertex
attribute names. Each igraph object in the list must have the time
period attribute called "period", a numeric or interger attribute.
Vertex coloring is automatically generated from rainbow pallette.
}
\author{
Stephen Downing
}

