\name{network-package}
\alias{network-package}
\alias{network}
\docType{package}
\title{
network
}
\description{
Wrapper functions and related helper functions to produce multiframe plots of igraph objects faceted by a vertex attribute and/or edge period attribute.
}
\details{
\tabular{ll}{
Package: \tab network\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-23\cr
License: \tab GPL (v3 or later)\cr
}
This package contains fuctions to make visual comparison and presentation of a single or multigrame graph plots facted by time period. Plotting functions include \code{multigraph()} for multiple graph plots with vertices contracted by a given attribute and \code{multinetwork()} for multiframe plots of the full networks.
}
\author{
Stephen Downing

Maintainer: Stephen Downing <sdowning.bm02g@nctu.edu.tw>
}
\keyword{ package }
\seealso{
\code{\link[igraph:igraph-package]{igraph}}
}
\examples{
coming soon...

}


%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~

